+++ Magic/CHANGELOG.md
@@ -25,6 +25,7 @@
  - Add carshop to war configs, and single-use wand items for spawning cars
  - Players are invisible and untargetable while riding in a car
  - /mmob clear will also clear non-mage mobs (like cars), though only in loaded chunks
+ - Add MobArena integration (PR to MobArena pending), use "magic:wand" in MobArena configs for Magic items
 
 ## 6.9.8